// types/job.d.ts
export type Job =
  | 'NONE'
  | 'GLA'
  | 'PGL'
  | 'MRD'
  | 'LNC'
  | 'ARC'
  | 'CNJ'
  | 'THM'
  | 'CRP'
  | 'BSM'
  | 'ARM'
  | 'GSM'
  | 'LTW'
  | 'WVR'
  | 'ALC'
  | 'CUL'
  | 'MIN'
  | 'BTN'
  | 'FSH'
  | 'PLD'
  | 'MNK'
  | 'WAR'
  | 'DRG'
  | 'BRD'
  | 'WHM'
  | 'BLM'
  | 'ACN'
  | 'SMN'
  | 'SCH'
  | 'ROG'
  | 'NIN'
  | 'MCH'
  | 'DRK'
  | 'AST'
  | 'SAM'
  | 'RDM'
  | 'BLU'
  | 'GNB'
  | 'DNC'
  | 'RPR'
  | 'SGE'
  | 'VPR'
  | 'PCT'
  | 'BST';

export type Role = 'tank' | 'healer' | 'dps' | 'crafter' | 'gatherer' | 'none';